+++ projProblox/Source/projProblox/Pickupables/Cores/CubeSocketInfo.cpp
@@ -87,7 +87,7 @@ TArray<FName> UCubeSocketInfo::GetFreeSockets() const
 {
 	TArray<FName> free;
 
-	for(int i = 0; i < socketObjects.Num(); i++)
+	for(unsigned short i = 0; i < socketObjects.Num(); i++)
 	{
 		if(IsValid(socketObjects[i])) continue;;
 
@@ -103,7 +103,7 @@ TArray<FName> UCubeSocketInfo::GetOccupiedSockets() const
 	if(socketObjects.IsEmpty()) return {};
 	
 	TArray<FName> output;
-	for (int i = 0; i < socketObjects.Num(); i++)
+	for (unsigned short i = 0; i < socketObjects.Num(); i++)
 	{
 		if(!IsValid(socketObjects[i])) continue;
